REFACTOR: Decomposed storage limit control code into multiple files and grouped into folder
